Quick extension to make $wgDebugComments useful.
Makes a click through on the purge page instead of redirect, since
that's where all the interesting debug messages are.

(Git makes it hard to work with locally modified files, so
turned it into an extension. Putting on svn because i just
want somewhere to stuff it, and i think there is some
rigamarole with getting a new extension in git. Its
a pretty useless extension, so I'm ok with it being on svn).